+++ pkg/operations/filter.go
